Shufling helper functions into their places
